Default view did not match homepage page part
I just installed Spina locally and noticed that the default homepage's text field was not rendering. After a bit of digging, it turns out that the page part name in the default.rb was changed some time ago, but the accessor variable used in the default homepage view was not updated.

Please don't mind my poor ability to describe technical ruby terms; I'm fairly new to this.